backport from trunk (#5648 + #5649): - fixed several compile wornings on BSD   Many thanks to Mark R for providing access to a BSD machine
- removed seas module from default compiling as it uses the dprintf function which is not available on Solaris or BSD


git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/branches/1.5@5650 689a6050-402a-0410-94f2-e92a70836424

+++ modules/ratelimit/ratelimit.c
@@ -57,7 +57,7 @@ MODULE_VERSION
  */
 #define RL_TIMER_INTERVAL 10
 
-#define RXLS(m, str, i) (m)[i].rm_eo - (m)[i].rm_so, (str) + (m)[i].rm_so
+#define RXLS(m, str, i) (int)((m)[i].rm_eo - (m)[i].rm_so),(str) + (m)[i].rm_so
 #define RXL(m, str, i) (m)[i].rm_eo - (m)[i].rm_so
 #define RXS(m, str, i) (str) + (m)[i].rm_so
